+++ app/src/free/AndroidManifest.xml
@@ -16,29 +16,24 @@
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
-  <application
-      android:name="com.jrummyapps.busybox.MainApp"
-      android:allowBackup="true"
-      android:label="@string/app_name"
-      android:supportsRtl="true">
-    <activity
-        android:name="com.jrummy.busybox.installer.activities.MainActivity"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/activity_busybox_installer"
-        android:launchMode="singleInstance"
-        android:theme="@style/Radiant.Light.DarkActionBar.NoActionBar">
-      <intent-filter>
-        <action android:name="android.intent.action.MAIN"/>
+    <application
+        android:name="com.jrummyapps.busybox.MainApp"
+        android:allowBackup="true"
+        android:label="@string/app_name"
+        android:supportsRtl="true"
+        >
+        <activity
+            android:name="com.jrummy.busybox.installer.activities.MainActivity"
+            android:icon="@mipmap/ic_launcher"
+            android:label="@string/activity_busybox_installer"
+            android:launchMode="singleInstance"
+            android:theme="@style/Radiant.Light.DarkActionBar.NoActionBar"
+            >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
 
-        <category android:name="android.intent.category.LAUNCHER"/>
-      </intent-filter>
-    </activity>
-
-    <service android:name="com.jrummy.busybox.installer.MyFirebaseMessagingService">
-      <intent-filter>
-        <action android:name="com.google.firebase.MESSAGING_EVENT" />
-      </intent-filter>
-    </service>
-  </application>
-
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+    </application>
 </manifest>
