+++ app/src/main/AndroidManifest.xml
@@ -91,6 +91,13 @@
 
     <activity android:name="com.jrummyapps.busybox.activities.CrossPromoActivity" />
 
+
+    <service android:name=".utils.MyFirebaseMessagingService">
+      <intent-filter>
+        <action android:name="com.google.firebase.MESSAGING_EVENT" />
+      </intent-filter>
+    </service>
+
     <receiver
         android:name="com.jrummyapps.busybox.InstalledEventReceiver"
         android:exported="false" />
