+++ app/src/paid/java/com/jrummy/busybox/installer/activities/MainActivity.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2016 Jared Rummler <jared.rummler@gmail.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.jrummy.busybox.installer.activities;
+
+
+import android.content.Context;
+import android.content.Intent;
+
+public class MainActivity extends com.jrummyapps.busybox.activities.MainActivity {
+    public static Intent linkIntent(Context context, String link) {
+        return new Intent(context, MainActivity.class)
+            .addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+            .putExtra(EXTRA_URI_KEY, link);
+    }
+}
