+++ app/src/staging/AndroidManifest.xml
@@ -20,25 +20,20 @@
       android:name="com.jrummyapps.busybox.MainApp"
       android:allowBackup="true"
       android:label="@string/app_name"
-      android:supportsRtl="true">
+      android:supportsRtl="true"
+      >
     <activity
-        android:name="com.jrummy.busybox.installer.MainActivity"
+        android:name="com.jrummy.busybox.installer.activities.MainActivity"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/activity_busybox_installer"
         android:launchMode="singleInstance"
-        android:theme="@style/Radiant.Light.DarkActionBar.NoActionBar">
+        android:theme="@style/Radiant.Light.DarkActionBar.NoActionBar"
+        >
       <intent-filter>
-        <action android:name="android.intent.action.MAIN"/>
+        <action android:name="android.intent.action.MAIN" />
 
-        <category android:name="android.intent.category.LAUNCHER"/>
+        <category android:name="android.intent.category.LAUNCHER" />
       </intent-filter>
     </activity>
-
-    <service android:name="com.jrummy.busybox.installer.MyFirebaseMessagingService">
-      <intent-filter>
-        <action android:name="com.google.firebase.MESSAGING_EVENT" />
-      </intent-filter>
-    </service>
   </application>
-
 </manifest>
